chore: refactor and remove redundant code
