+++ Cargo.toml
@@ -2,7 +2,6 @@
 name = "fossil_headers_db"
 version = "0.1.0"
 edition = "2021"
-default-run = "fossil_headers_db"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -35,11 +34,6 @@ serde_json = "1.0.105"
 serial_test = "3.2.0"
 testcontainers-modules = { version = "0.11.5", features = ["postgres"] }
 
-
-[[bin]]
-name = "fossil_headers_db"
-path = "src/main.rs"
-
 [[bin]]
 name = "fossil_indexer"
-path = "src/indexer/main.rs"
+path = "src/main.rs"
