--- Dockerfile.fix
@@ -1,21 +1,0 @@
-FROM rust:1.81 AS builder
-
-WORKDIR /usr/app
-COPY . .
-RUN cargo build --release
-
-FROM debian:bookworm-slim
-
-# Install necessary runtime dependencies
-RUN apt-get update && \
-    apt-get install -y --no-install-recommends ca-certificates && \
-    rm -rf /var/lib/apt/lists/*
-
-WORKDIR /usr/app
-
-# Copy the compiled binary
-COPY --from=builder /usr/app/target/release/fossil_headers_db .
-
-EXPOSE 8080
-
-CMD ["/usr/app/fossil_headers_db", "fix"]
