+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub mod commands;
 pub mod db;
 pub mod indexer;
 pub mod repositories;
