+++ src/rpc/mod.rs
@@ -1,23 +1,9 @@
 use crate::utils::convert_hex_string_to_i64;
 use eyre::{eyre, Context, Result};
-use once_cell::sync::Lazy;
-use reqwest::Client;
 use serde::{Deserialize, Serialize};
 use std::{future::Future, time::Duration};
 use tokio::time::sleep;
 use tracing::{error, warn};
-
-// TODO: instead of keeping this static, make it passable as a dependency.
-// This should allow us to test this module.
-static CLIENT: Lazy<Client> = Lazy::new(Client::new);
-static NODE_CONNECTION_STRING: Lazy<Option<String>> = Lazy::new(|| {
-    dotenvy::var("NODE_CONNECTION_STRING")
-        .map_err(|e| error!("Failed to get NODE_CONNECTION_STRING: {}", e))
-        .ok()
-});
-
-// Arbitrarily set, can be set somewhere else.
-const MAX_RETRIES: u8 = 5;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct RpcResponse<T> {
@@ -53,234 +39,6 @@ pub struct Transaction {
     pub chain_id: Option<String>,
 }
 
-#[derive(Debug, Deserialize, Clone)]
-#[allow(dead_code)]
-pub struct BlockHeaderWithEmptyTransaction {
-    #[serde(rename(deserialize = "gasLimit"))]
-    pub gas_limit: String,
-    #[serde(rename(deserialize = "gasUsed"))]
-    pub gas_used: String,
-    #[serde(rename(deserialize = "baseFeePerGas"))]
-    pub base_fee_per_gas: Option<String>,
-    pub hash: String,
-    pub nonce: Option<String>,
-    pub number: String,
-    #[serde(rename(deserialize = "receiptsRoot"))]
-    pub receipts_root: Option<String>,
-    #[serde(rename(deserialize = "stateRoot"))]
-    pub state_root: Option<String>,
-    #[serde(rename(deserialize = "transactionsRoot"))]
-    pub transactions_root: Option<String>,
-    #[serde(rename(deserialize = "parentHash"))]
-    pub parent_hash: Option<String>,
-    #[serde(rename(deserialize = "miner"))]
-    pub miner: Option<String>,
-    #[serde(rename(deserialize = "logsBloom"))]
-    pub logs_bloom: Option<String>,
-    #[serde(rename(deserialize = "difficulty"))]
-    pub difficulty: Option<String>,
-    #[serde(rename(deserialize = "totalDifficulty"))]
-    pub total_difficulty: Option<String>,
-    #[serde(rename(deserialize = "sha3Uncles"))]
-    pub sha3_uncles: Option<String>,
-    #[serde(rename(deserialize = "timestamp"))]
-    pub timestamp: String,
-    #[serde(rename(deserialize = "extraData"))]
-    pub extra_data: Option<String>,
-    #[serde(rename(deserialize = "mixHash"))]
-    pub mix_hash: Option<String>,
-    #[serde(rename(deserialize = "withdrawalsRoot"))]
-    pub withdrawals_root: Option<String>,
-    #[serde(rename(deserialize = "blobGasUsed"))]
-    pub blob_gas_used: Option<String>,
-    #[serde(rename(deserialize = "excessBlobGas"))]
-    pub excess_blob_gas: Option<String>,
-    #[serde(rename(deserialize = "parentBeaconBlockRoot"))]
-    pub parent_beacon_block_root: Option<String>,
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct BlockHeaderWithFullTransaction {
-    #[serde(rename(deserialize = "gasLimit"))]
-    pub gas_limit: String,
-    #[serde(rename(deserialize = "gasUsed"))]
-    pub gas_used: String,
-    #[serde(rename(deserialize = "baseFeePerGas"))]
-    pub base_fee_per_gas: Option<String>,
-    pub hash: String,
-    pub nonce: Option<String>,
-    pub number: String,
-    #[serde(rename(deserialize = "receiptsRoot"))]
-    pub receipts_root: Option<String>,
-    #[serde(rename(deserialize = "stateRoot"))]
-    pub state_root: Option<String>,
-    #[serde(rename(deserialize = "transactionsRoot"))]
-    pub transactions_root: Option<String>,
-    pub transactions: Vec<Transaction>,
-    #[serde(rename(deserialize = "parentHash"))]
-    pub parent_hash: Option<String>,
-    pub miner: Option<String>,
-    #[serde(rename(deserialize = "logsBloom"))]
-    pub logs_bloom: Option<String>,
-    #[serde(rename(deserialize = "difficulty"))]
-    pub difficulty: Option<String>,
-    #[serde(rename(deserialize = "totalDifficulty"))]
-    pub total_difficulty: Option<String>,
-    #[serde(rename(deserialize = "sha3Uncles"))]
-    pub sha3_uncles: Option<String>,
-    #[serde(rename(deserialize = "timestamp"))]
-    pub timestamp: String,
-    #[serde(rename(deserialize = "extraData"))]
-    pub extra_data: Option<String>,
-    #[serde(rename(deserialize = "mixHash"))]
-    pub mix_hash: Option<String>,
-    #[serde(rename(deserialize = "withdrawalsRoot"))]
-    pub withdrawals_root: Option<String>,
-    #[serde(rename(deserialize = "blobGasUsed"))]
-    pub blob_gas_used: Option<String>,
-    #[serde(rename(deserialize = "excessBlobGas"))]
-    pub excess_blob_gas: Option<String>,
-    #[serde(rename(deserialize = "parentBeaconBlockRoot"))]
-    pub parent_beacon_block_root: Option<String>,
-}
-
-pub async fn get_latest_finalized_blocknumber(timeout: Option<u64>) -> Result<i64> {
-    // TODO: Id should be different on every request, this is how request are identified by us and by the node.
-    let params = RpcRequest {
-        jsonrpc: "2.0",
-        id: "0".to_string(),
-        method: "eth_getBlockByNumber",
-        params: ("finalized", false),
-    };
-
-    match make_retrying_rpc_call::<_, BlockHeaderWithEmptyTransaction>(
-        &params,
-        timeout,
-        MAX_RETRIES.into(),
-    )
-    .await
-    .context("Failed to get latest block number")
-    {
-        Ok(blockheader) => Ok(convert_hex_string_to_i64(&blockheader.number)?),
-        Err(e) => Err(e),
-    }
-}
-
-pub async fn get_full_block_by_number(
-    number: i64,
-    timeout: Option<u64>,
-) -> Result<BlockHeaderWithFullTransaction> {
-    let params = RpcRequest {
-        jsonrpc: "2.0",
-        id: "0".to_string(),
-        method: "eth_getBlockByNumber",
-        params: (format!("0x{:x}", number), true),
-    };
-
-    make_retrying_rpc_call::<_, BlockHeaderWithFullTransaction>(
-        &params,
-        timeout,
-        MAX_RETRIES.into(),
-    )
-    .await
-}
-
-// TODO: Make this work as expected
-#[allow(dead_code)]
-pub async fn batch_get_full_block_by_number(
-    numbers: Vec<i64>,
-    timeout: Option<u64>,
-) -> Result<Vec<BlockHeaderWithFullTransaction>> {
-    let mut params = Vec::new();
-    for number in numbers {
-        let num_str = number.to_string();
-        params.push(RpcRequest {
-            jsonrpc: "2.0",
-            id: num_str,
-            method: "eth_getBlockByNumber",
-            params: (format!("0x{:x}", number), true),
-        });
-    }
-    make_rpc_call::<_, Vec<BlockHeaderWithFullTransaction>>(&params, timeout).await
-}
-
-async fn make_rpc_call<T: Serialize, R: for<'de> Deserialize<'de>>(
-    params: &T,
-    timeout: Option<u64>,
-) -> Result<R> {
-    let connection_string = (*NODE_CONNECTION_STRING)
-        .as_ref()
-        .ok_or_else(|| eyre::eyre!("NODE_CONNECTION_STRING not set"))?;
-
-    let raw_response = match timeout {
-        Some(seconds) => {
-            CLIENT
-                .post(connection_string)
-                .timeout(Duration::from_secs(seconds))
-                .json(params)
-                .send()
-                .await
-        }
-        None => CLIENT.post(connection_string).json(params).send().await,
-    };
-
-    let raw_response = match raw_response {
-        Ok(response) => response,
-        Err(e) => {
-            error!("HTTP request error: {:?}", e);
-            return Err(e.into());
-        }
-    };
-
-    // Attempt to extract JSON from the response
-    let json_response = raw_response.text().await;
-    match json_response {
-        Ok(text) => {
-            // Try to deserialize the response, logging if it fails
-            match serde_json::from_str::<RpcResponse<R>>(&text) {
-                Ok(parsed) => Ok(parsed.result),
-                Err(e) => {
-                    error!(
-                        "Deserialization error: {:?}\nResponse snippet: {:?}",
-                        e,
-                        text // Log the entire response
-                    );
-                    Err(e.into())
-                }
-            }
-        }
-        Err(e) => {
-            error!("Failed to read response body: {:?}", e);
-            Err(e.into())
-        }
-    }
-}
-
-async fn make_retrying_rpc_call<T: Serialize, R: for<'de> Deserialize<'de> + Send>(
-    params: &T,
-    timeout: Option<u64>,
-    max_retries: u32,
-) -> Result<R> {
-    let mut attempts = 0;
-    loop {
-        match make_rpc_call(params, timeout).await {
-            Ok(result) => return Ok(result),
-            Err(e) => {
-                attempts += 1;
-                if attempts > max_retries {
-                    warn!("Operation failed with error: {:?}. Max retries reached", e);
-                    return Err(e);
-                }
-                let backoff = Duration::from_secs(2_u64.pow(attempts));
-                warn!(
-                    "Operation failed with error: {:?}. Retrying in {:?} (Attempt {}/{})",
-                    e, backoff, attempts, max_retries
-                );
-                sleep(backoff).await;
-            }
-        }
-    }
-}
 #[derive(Debug, Serialize, Deserialize, Clone)]
 #[serde(untagged)]
 pub enum BlockTransaction {
@@ -344,7 +102,6 @@ pub struct BlockHeader {
     pub parent_beacon_block_root: Option<String>,
 }
 
-#[allow(dead_code)]
 pub trait EthereumRpcProvider {
     fn get_latest_finalized_blocknumber(
         &self,
@@ -366,7 +123,6 @@ pub struct EthereumJsonRpcClient {
 }
 
 impl EthereumJsonRpcClient {
-    #[allow(dead_code)]
     pub fn new(connection_string: String, max_retries: u32) -> Self {
         Self {
             client: reqwest::Client::new(),
@@ -512,6 +268,24 @@ pub fn try_convert_full_tx_vector(block_tx_vec: Vec<BlockTransaction>) -> Result
 
     Ok(result_vec)
 }
+
+// TODO: Make this work as expected
+// pub async fn batch_get_full_block_by_number(
+//     numbers: Vec<i64>,
+//     timeout: Option<u64>,
+// ) -> Result<Vec<BlockHeaderWithFullTransaction>> {
+//     let mut params = Vec::new();
+//     for number in numbers {
+//         let num_str = number.to_string();
+//         params.push(RpcRequest {
+//             jsonrpc: "2.0",
+//             id: num_str,
+//             method: "eth_getBlockByNumber",
+//             params: (format!("0x{:x}", number), true),
+//         });
+//     }
+//     make_rpc_call::<_, Vec<BlockHeaderWithFullTransaction>>(&params, timeout).await
+// }
 
 /// BIG TODO:
 /// Handle rpc errors correctly!
